refactor: extract table binding into separate class
